Presentation: Fama and French (1992)
These slides are my midterm presentation about Fama and French (1992) in the Intro to Finance Research course. These slides summarize the method Fama and French used to find factors that can explain the expected returns in the US stock market. This paper gave a solid ground for the three-factor and subsequent works exploring the "factors" role in asset pricing.