Replicated FF 1992 paper using TWSE data
In these slides, we first briefly review the basic concepts of Fama and French (1992), then we show the empirical results of the replication of this paper (using Taiwan Stock Exchange data from the Taiwan Economic Journal).  The results can be summarized as follows:

Our empirical results show that the size effect can strongly explain the cross-section of expected stock returns of companies listed on TWSE during 2005/7-2020/6. Tough, the effect of the BM ratio is positive, the result seems not to be reliable due to its low t-statistics. Moreover, under any combination, β does not explain the cross-section of expected stock returns in Taiwan during 2005/7-2020/6.