Main Article
This article is my final proposal for the Intro to Finance course. The article presents the results of the replication of Fama and French (1922) using data from the Taiwan stock market as shown in the slides (1992rep.pdf). Moreover, we also conduct a primary investigation of these factors at the industry level. We conclude that there might exist heterogeneous behaviors at the industry level and call for further research.